services/api/requirements.txt
uvicorn==0.29.0
python-dotenv==1.0.1
fastapi==0.111.0
pydantic==2.7.1
fastapi-route-logger-middleware==0.2.3
